+++ tesseract_collision/test/benchmarks/collision_profile.cpp
@@ -14,6 +14,7 @@ TESSERACT_COMMON_IGNORE_WARNINGS_POP
 #include <tesseract_collision/hpp_fcl/hpp_fcl_discrete_managers.h>
 
 #include <tesseract_geometry/impl/sphere.h>
+#include <tesseract_common/resource_locator.h>
 
 static const std::size_t DIM = 10;
 
@@ -32,8 +33,12 @@ void addCollisionObjects(DiscreteContactManager& checker, bool use_single_link, 
   {
     auto mesh_vertices = std::make_shared<tesseract_common::VectorVector3d>();
     auto mesh_faces = std::make_shared<Eigen::VectorXi>();
-    loadSimplePlyFile(
-        std::string(TESSERACT_SUPPORT_DIR) + "/meshes/sphere_p25m.ply", *mesh_vertices, *mesh_faces, true);
+
+    tesseract_common::GeneralResourceLocator locator;
+    loadSimplePlyFile(locator.locateResource("package://tesseract_support/meshes/sphere_p25m.ply")->getFilePath(),
+                      *mesh_vertices,
+                      *mesh_faces,
+                      true);
 
     auto mesh = std::make_shared<tesseract_geometry::Mesh>(mesh_vertices, mesh_faces);
     sphere = makeConvexMesh(*mesh);
@@ -99,8 +104,12 @@ void addCollisionObjects(ContinuousContactManager& checker, bool use_single_link
   {
     auto mesh_vertices = std::make_shared<tesseract_common::VectorVector3d>();
     auto mesh_faces = std::make_shared<Eigen::VectorXi>();
-    loadSimplePlyFile(
-        std::string(TESSERACT_SUPPORT_DIR) + "/meshes/sphere_p25m.ply", *mesh_vertices, *mesh_faces, true);
+
+    tesseract_common::GeneralResourceLocator locator;
+    loadSimplePlyFile(locator.locateResource("package://tesseract_support/meshes/sphere_p25m.ply")->getFilePath(),
+                      *mesh_vertices,
+                      *mesh_faces,
+                      true);
 
     auto mesh = std::make_shared<tesseract_geometry::Mesh>(mesh_vertices, mesh_faces);
     sphere = makeConvexMesh(*mesh);