+++ tesseract_collision/test_suite/include/tesseract_collision/test_suite/collision_box_sphere_unit.hpp
@@ -5,6 +5,7 @@
 #include <tesseract_collision/core/discrete_contact_manager.h>
 #include <tesseract_collision/core/common.h>
 #include <tesseract_geometry/geometries.h>
+#include <tesseract_common/resource_locator.h>
 
 namespace tesseract_collision::test_suite
 {
@@ -52,8 +53,12 @@ inline void addCollisionObjects(DiscreteContactManager& checker, bool use_convex
   {
     auto mesh_vertices = std::make_shared<tesseract_common::VectorVector3d>();
     auto mesh_faces = std::make_shared<Eigen::VectorXi>();
-    EXPECT_GT(loadSimplePlyFile(
-                  std::string(TESSERACT_SUPPORT_DIR) + "/meshes/sphere_p25m.ply", *mesh_vertices, *mesh_faces, true),
+    tesseract_common::GeneralResourceLocator locator;
+    EXPECT_GT(
+        loadSimplePlyFile(locator.locateResource("package://tesseract_support/meshes/sphere_p25m.ply")->getFilePath(),
+                          *mesh_vertices,
+                          *mesh_faces,
+                          true),
               0);
 
     auto mesh = std::make_shared<tesseract_geometry::Mesh>(mesh_vertices, mesh_faces);
@@ -79,7 +84,7 @@ inline void addCollisionObjects(DiscreteContactManager& checker, bool use_convex
   /////////////////////////////////////////////
   CollisionShapePtr remove_box = std::make_shared<tesseract_geometry::Box>(0.1, 1, 1);
   Eigen::Isometry3d remove_box_pose;
-  thin_box_pose.setIdentity();
+  remove_box_pose.setIdentity();
 
   CollisionShapesConst obj4_shapes;
   tesseract_common::VectorIsometry3d obj4_poses;